+++ src/ECOFORGE/VISTA/PanelVenta.form
@@ -320,7 +320,6 @@
                     <StringItem index="2" value="A Cuotas"/>
                   </StringArray>
                 </Property>
-                <Property name="selectedIndex" type="int" value="0"/>
               </Properties>
               <AuxValues>
                 <AuxValue name="JavaCodeGenerator_TypeParameters" type="java.lang.String" value="&lt;String&gt;"/>
