--- src/ECOFORGE/VISTA/datosProyecto.java
@@ -1,192 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package ECOFORGE.VISTA;
-
-/**
- *
- * @author YANFER
- */
-import ECOFORGE.CONTROLADOR.ControladorProyecto;
-import ECOFORGE.MODELO.Proyecto;
-import javax.swing.JOptionPane;
-
-public class datosProyecto extends javax.swing.JFrame {
-
-    private ControladorProyecto controlador;
-    private boolean isModoActualizar = false;
-    private Proyectos formularioProyecto;
-
-    /**
-     * Creates new form datosProyecto
-     */
-    public datosProyecto(Proyectos formularioProyecto) {
-        initComponents();
-        this.formularioProyecto = formularioProyecto;
-        controlador = new ControladorProyecto();
-        configurarFormulario();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanel1 = new javax.swing.JPanel();
-        jTextFieldCodProyecto = new javax.swing.JTextField();
-        jTextFieldNombreProyecto = new javax.swing.JTextField();
-        jLabel1 = new javax.swing.JLabel();
-        jLabel2 = new javax.swing.JLabel();
-        jButton1 = new javax.swing.JButton();
-        jButton2 = new javax.swing.JButton();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setTitle("Datos del Proyecto");
-        getContentPane().setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        jPanel1.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-        getContentPane().add(jPanel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(544, 92, -1, -1));
-        getContentPane().add(jTextFieldCodProyecto, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 100, 171, -1));
-        getContentPane().add(jTextFieldNombreProyecto, new org.netbeans.lib.awtextra.AbsoluteConstraints(400, 100, 192, -1));
-
-        jLabel1.setText("Codigo");
-        getContentPane().add(jLabel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 80, 46, -1));
-
-        jLabel2.setText("Nombre");
-        getContentPane().add(jLabel2, new org.netbeans.lib.awtextra.AbsoluteConstraints(400, 80, 102, -1));
-
-        jButton1.setText("Guardar");
-        jButton1.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton1ActionPerformed(evt);
-            }
-        });
-        getContentPane().add(jButton1, new org.netbeans.lib.awtextra.AbsoluteConstraints(220, 300, 75, -1));
-
-        jButton2.setText("Actualizar");
-        jButton2.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jButton2ActionPerformed(evt);
-            }
-        });
-        getContentPane().add(jButton2, new org.netbeans.lib.awtextra.AbsoluteConstraints(430, 300, 99, -1));
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-    
-    private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
-        // Obtener datos de los campos de texto
-        Integer codigo_proyecto = Integer.parseInt(jTextFieldCodProyecto.getText());
-        String nombre_proyecto = jTextFieldNombreProyecto.getText();
-
-        // Crear un objeto Proyecto
-        Proyecto nuevoProyecto = new Proyecto(codigo_proyecto, nombre_proyecto);
-
-        // Agregar Proyecto utilizando el controlador
-        if (controlador.crearProyecto(nuevoProyecto)) {
-            JOptionPane.showMessageDialog(this, "Proyecto agregado exitosamente.");
-
-            // Actualizar la tabla en el formulario Proyecto
-            formularioProyecto.actualizarTabla();
-
-            this.dispose(); // Cerrar el formulario
-        } else {
-            JOptionPane.showMessageDialog(this, "Error al agregar Proyecto.");
-        }
-    }//GEN-LAST:event_jButton1ActionPerformed
-
-    private void jButton2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton2ActionPerformed
-        if (isModoActualizar) {
-            // Código para actualizar el Proyecto
-            Integer codigo_proyecto = Integer.parseInt(jTextFieldCodProyecto.getText());
-            String nombre_proyecto = jTextFieldNombreProyecto.getText();
-            Proyecto actualizarProyecto = new Proyecto(codigo_proyecto, nombre_proyecto);
-            // Llamar al controlador para actualizar
-            if (controlador.actualizarProyecto(actualizarProyecto)) {
-                JOptionPane.showMessageDialog(this, "Proyecto actualizado correctamente.");
-                // Cerrar el formulario después de actualizar
-                formularioProyecto.actualizarTabla();
-                this.dispose();
-            } else {
-                JOptionPane.showMessageDialog(this, "Error al actualizar el Proyecto.");
-            }
-        } else {
-            // Código para agregar un nuevo Proyecto
-            // (Ya lo tienes implementado)
-        }        // TODO add your handling code here:
-    }//GEN-LAST:event_jButton2ActionPerformed
-    public void cargarDatos(Integer codigo_proyecto, String nombre_proyecto) {
-        jTextFieldCodProyecto.setText(String.valueOf(codigo_proyecto));
-        jTextFieldNombreProyecto.setText(nombre_proyecto);
-
-        // Deshabilitar la caja de texto de cédula para que no sea modificable
-        jTextFieldCodProyecto.setEnabled(false);
-    }
-
-    public void setModoActualizar(boolean modoActualizar) {
-        this.isModoActualizar = modoActualizar;
-    }
-    
-    private void configurarFormulario() {
-        // jButton2.setEnabled(true); // Desactivar el botón de Actualizar
-        // jButton1.setEnabled(true);     // Activar el botón de Agregar
-        if (isModoActualizar) {
-            jButton2.setEnabled(true);  // Activar el botón de Actualizar
-            jButton1.setEnabled(false);    // Desactivar el botón de Agregar
-        } else {
-            jButton2.setEnabled(false); // Desactivar el botón de Actualizar
-            jButton1.setEnabled(true);     // Activar el botón de Agregar
-        }
-    }
-    
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(datosProyecto.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(datosProyecto.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(datosProyecto.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(datosProyecto.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new datosProyecto(new Proyectos()).setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton jButton1;
-    private javax.swing.JButton jButton2;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JTextField jTextFieldCodProyecto;
-    private javax.swing.JTextField jTextFieldNombreProyecto;
-    // End of variables declaration//GEN-END:variables
-}
